requirements.txt
Django==5.0.4
django-allauth[socialaccount]==0.62.1
django-ninja==1.1.0
whitenoise==6.6.0
psycopg2-binary==2.9.10
python-decouple==3.8
Pillow==10.4.0
gunicorn==21.2.0
redis==5.0.4
django-redis==5.4.0
